feat: create shared types module to eliminate three-way duplication
Create a new shared/types/ directory with single source of truth for
all type definitions used across main process, renderer process, and
preload script.

New structure:
- shared/types/domain.ts - Core business entities (Worktree, DevServerState, etc.)
- shared/types/ipc.ts - IPC payloads and ElectronAPI interface
- shared/types/config.ts - Application configuration types
- shared/types/keymap.ts - Keyboard shortcut types
- shared/types/index.ts - Barrel file re-exporting all types

Benefits:
- Single source of truth eliminates sync issues between processes
- ~275 lines removed from codebase through consolidation
- Type changes now propagate automatically across all consumers
- Better IDE support with unified type definitions

Updated imports across all consumers:
- electron/types/* now re-exports from shared (keeping runtime values local)
- electron/preload.ts reduced from ~250 inline types to clean imports
- src/types/* now re-exports from shared
- All component and store imports updated

Fixes #87
