+++ electron/preload.ts
@@ -10,13 +10,41 @@
  * - Listeners provide cleanup functions to prevent memory leaks
  *
  * NOTE: This file is built separately as CommonJS for Electron's sandboxed preload.
- * Channel names and types are inlined to avoid module format conflicts with the
- * ESM main process.
+ * Types are imported from the shared module but channel names are inlined to avoid
+ * module format conflicts with the ESM main process.
  */
 
 import { contextBridge, ipcRenderer } from "electron";
 
+// Import types from shared module
+import type {
+  WorktreeState,
+  DevServerState,
+  Project,
+  TerminalSpawnOptions,
+  CopyTreeOptions,
+  CopyTreeResult,
+  CopyTreeProgress,
+  AppState,
+  LogEntry,
+  LogFilterOptions,
+  EventRecord,
+  EventFilterOptions,
+  RetryAction,
+  AppError,
+  AgentSession,
+  HistoryGetSessionsPayload,
+  AIServiceState,
+  ProjectIdentity,
+  AgentStateChangePayload,
+  ElectronAPI,
+} from "@shared/types";
+
+// Re-export ElectronAPI for type declarations
+export type { ElectronAPI };
+
 // Inlined channel constants (must match electron/ipc/channels.ts)
+// These are kept inline to avoid runtime module resolution issues with CommonJS
 const CHANNELS = {
   // Worktree channels
   WORKTREE_GET_ALL: "worktree:get-all",
@@ -121,399 +149,6 @@ const CHANNELS = {
   AI_GENERATE_PROJECT_IDENTITY: "ai:generate-project-identity",
 } as const;
 
-// Inlined types (must match electron/ipc/types.ts)
-interface WorktreeState {
-  id: string;
-  path: string;
-  name: string;
-  branch?: string;
-  isCurrent: boolean;
-  isMainWorktree: boolean;
-  gitDir: string;
-  summary?: string;
-  modifiedCount?: number;
-  mood?: "active" | "stable" | "stale" | "error";
-  aiStatus?: "active" | "loading" | "disabled" | "error";
-  lastActivityTimestamp?: number;
-  aiNote?: string;
-  aiNoteTimestamp?: number;
-  issueNumber?: number;
-  prNumber?: number;
-  prUrl?: string;
-  prState?: "open" | "merged" | "closed";
-  worktreeChanges?: WorktreeChanges | null;
-}
-
-interface WorktreeChanges {
-  worktreeId: string;
-  worktreePath: string;
-  files: FileChangeDetail[];
-  changedFileCount: number;
-  insertions: number;
-  deletions: number;
-  latestMtime: number;
-  timestamp: number;
-}
-
-interface FileChangeDetail {
-  path: string;
-  status: "added" | "modified" | "deleted" | "renamed" | "copied" | "untracked";
-  insertions: number;
-  deletions: number;
-  mtime?: number;
-}
-
-interface DevServerState {
-  worktreeId: string;
-  status: "stopped" | "starting" | "running" | "error";
-  url?: string;
-  port?: number;
-  pid?: number;
-  errorMessage?: string;
-  logs: string[];
-}
-
-interface TerminalSpawnOptions {
-  id?: string;
-  cwd: string;
-  shell?: string;
-  args?: string[];
-  env?: Record<string, string>;
-  cols?: number;
-  rows?: number;
-  type?: "shell" | "claude" | "gemini" | "custom";
-  title?: string;
-  worktreeId?: string;
-}
-
-interface CopyTreeOptions {
-  /** Output format */
-  format?: "xml" | "json" | "markdown" | "tree" | "ndjson";
-
-  /** Pattern filtering */
-  filter?: string | string[];
-  exclude?: string | string[];
-  always?: string[];
-
-  /** Git filtering */
-  modified?: boolean;
-  changed?: string;
-
-  /** Size limits */
-  maxFileSize?: number;
-  maxTotalSize?: number;
-  maxFileCount?: number;
-
-  /** Formatting */
-  withLineNumbers?: boolean;
-  charLimit?: number;
-
-  /** Profile (load from .copytree file) - legacy option */
-  profile?: string;
-}
-
-interface CopyTreeResult {
-  content: string;
-  fileCount: number;
-  error?: string;
-  stats?: {
-    totalSize: number;
-    duration: number;
-  };
-}
-
-interface CopyTreeProgress {
-  /** Current stage name (e.g., 'FileDiscoveryStage', 'FormatterStage') */
-  stage: string;
-  /** Progress percentage (0-1) */
-  progress: number;
-  /** Human-readable progress message */
-  message: string;
-  /** Files processed so far (if known) */
-  filesProcessed?: number;
-  /** Total files estimated (if known) */
-  totalFiles?: number;
-  /** Current file being processed (if known) */
-  currentFile?: string;
-}
-
-interface CanopyConfig {
-  editor?: string;
-  editorArgs?: string[];
-  theme?: "dark" | "light";
-  monitor?: {
-    pollIntervalActive?: number;
-    pollIntervalBackground?: number;
-  };
-  ai?: {
-    enabled?: boolean;
-    summaryDebounceMs?: number;
-  };
-  devServer?: {
-    enabled?: boolean;
-    autoStart?: boolean;
-    customCommands?: Record<string, string>;
-  };
-  quickLinks?: {
-    enabled?: boolean;
-    links?: Array<{ name: string; url: string }>;
-  };
-  copytree?: {
-    defaultProfile?: string;
-    extraArgs?: string[];
-  };
-  keymap?: {
-    preset?: "standard" | "vim";
-    overrides?: Record<string, string>;
-  };
-}
-
-interface TerminalState {
-  id: string;
-  type: "shell" | "claude" | "gemini" | "custom";
-  title: string;
-  cwd: string;
-  worktreeId?: string;
-}
-
-interface AppState {
-  rootPath?: string;
-  terminals: TerminalState[];
-}
-
-type LogLevel = "debug" | "info" | "warn" | "error";
-
-interface LogEntry {
-  id: string;
-  timestamp: number;
-  level: LogLevel;
-  message: string;
-  context?: Record<string, unknown>;
-  source?: string;
-}
-
-interface LogFilterOptions {
-  levels?: LogLevel[];
-  sources?: string[];
-  search?: string;
-  startTime?: number;
-  endTime?: number;
-}
-
-// Event Inspector types
-interface EventRecord {
-  id: string;
-  timestamp: number;
-  type: string;
-  payload: any;
-  source: "main" | "renderer";
-}
-
-interface EventFilterOptions {
-  types?: string[];
-  worktreeId?: string;
-  agentId?: string;
-  taskId?: string;
-  search?: string;
-  after?: number;
-  before?: number;
-}
-
-// Project types
-interface Project {
-  id: string;
-  path: string;
-  name: string;
-  emoji: string;
-  aiGeneratedName?: string;
-  aiGeneratedEmoji?: string;
-  lastOpened: number;
-  color?: string;
-}
-
-// Agent session and transcript types
-interface TranscriptEntry {
-  timestamp: number;
-  type: "user" | "agent" | "system";
-  content: string;
-}
-
-interface Artifact {
-  id: string;
-  type: "code" | "patch" | "file" | "summary" | "other";
-  language?: string;
-  filename?: string;
-  content: string;
-  extractedAt: number;
-}
-
-interface AgentSession {
-  id: string;
-  agentType: "claude" | "gemini" | "custom";
-  worktreeId?: string;
-  startTime: number;
-  endTime?: number;
-  state: "active" | "completed" | "failed";
-  transcript: TranscriptEntry[];
-  artifacts: Artifact[];
-  metadata: {
-    terminalId: string;
-    cwd: string;
-    exitCode?: number;
-  };
-}
-
-interface HistoryGetSessionsPayload {
-  worktreeId?: string;
-  agentType?: "claude" | "gemini" | "custom";
-  limit?: number;
-}
-
-// Error types for IPC
-type ErrorType = "git" | "process" | "filesystem" | "network" | "config" | "unknown";
-type RetryAction = "copytree" | "devserver" | "terminal" | "git" | "worktree";
-
-// AI types
-interface AIConfig {
-  hasKey: boolean;
-  model: string;
-  enabled: boolean;
-}
-
-interface ProjectIdentity {
-  emoji: string;
-  title: string;
-  gradientStart: string;
-  gradientEnd: string;
-}
-
-interface AppError {
-  id: string;
-  timestamp: number;
-  type: ErrorType;
-  message: string;
-  details?: string;
-  source?: string;
-  context?: {
-    worktreeId?: string;
-    terminalId?: string;
-    filePath?: string;
-    command?: string;
-  };
-  isTransient: boolean;
-  dismissed: boolean;
-  retryAction?: RetryAction;
-  retryArgs?: Record<string, unknown>;
-}
-
-export interface ElectronAPI {
-  worktree: {
-    getAll(): Promise<WorktreeState[]>;
-    refresh(): Promise<void>;
-    setActive(worktreeId: string): Promise<void>;
-    onUpdate(callback: (state: WorktreeState) => void): () => void;
-    onRemove(callback: (data: { worktreeId: string }) => void): () => void;
-  };
-  devServer: {
-    start(worktreeId: string, worktreePath: string, command?: string): Promise<DevServerState>;
-    stop(worktreeId: string): Promise<DevServerState>;
-    toggle(worktreeId: string, worktreePath: string, command?: string): Promise<DevServerState>;
-    getState(worktreeId: string): Promise<DevServerState>;
-    getLogs(worktreeId: string): Promise<string[]>;
-    hasDevScript(worktreePath: string): Promise<boolean>;
-    onUpdate(callback: (state: DevServerState) => void): () => void;
-    onError(callback: (data: { worktreeId: string; error: string }) => void): () => void;
-  };
-  terminal: {
-    spawn(options: TerminalSpawnOptions): Promise<string>;
-    write(id: string, data: string): void;
-    resize(id: string, cols: number, rows: number): void;
-    kill(id: string): Promise<void>;
-    onData(id: string, callback: (data: string) => void): () => void;
-    onExit(callback: (id: string, exitCode: number) => void): () => void;
-    onAgentStateChanged(
-      callback: (data: {
-        agentId: string;
-        state: string;
-        previousState?: string;
-        timestamp: number;
-      }) => void
-    ): () => void;
-  };
-  copyTree: {
-    generate(worktreeId: string, options?: CopyTreeOptions): Promise<CopyTreeResult>;
-    injectToTerminal(
-      terminalId: string,
-      worktreeId: string,
-      options?: CopyTreeOptions
-    ): Promise<CopyTreeResult>;
-    isAvailable(): Promise<boolean>;
-    cancel(): Promise<void>;
-    onProgress(callback: (progress: CopyTreeProgress) => void): () => void;
-  };
-  system: {
-    openExternal(url: string): Promise<void>;
-    openPath(path: string): Promise<void>;
-    getConfig(): Promise<CanopyConfig>;
-    checkCommand(command: string): Promise<boolean>;
-  };
-  app: {
-    getState(): Promise<AppState>;
-    setState(partialState: Partial<AppState>): Promise<void>;
-  };
-  logs: {
-    getAll(filters?: LogFilterOptions): Promise<LogEntry[]>;
-    getSources(): Promise<string[]>;
-    clear(): Promise<void>;
-    openFile(): Promise<void>;
-    onEntry(callback: (entry: LogEntry) => void): () => void;
-  };
-  directory: {
-    getRecent(): Promise<Array<{ path: string; lastOpened: number; name: string }>>;
-    open(path: string): Promise<void>;
-    openDialog(): Promise<string | null>;
-    removeRecent(path: string): Promise<void>;
-  };
-  errors: {
-    onError(callback: (error: AppError) => void): () => void;
-    retry(errorId: string, action: RetryAction, args?: Record<string, unknown>): Promise<void>;
-    openLogs(): Promise<void>;
-  };
-  eventInspector: {
-    getEvents(): Promise<EventRecord[]>;
-    getFiltered(filters: EventFilterOptions): Promise<EventRecord[]>;
-    clear(): Promise<void>;
-    subscribe(): void;
-    unsubscribe(): void;
-    onEvent(callback: (event: EventRecord) => void): () => void;
-  };
-  project: {
-    getAll(): Promise<Project[]>;
-    getCurrent(): Promise<Project | null>;
-    add(path: string): Promise<Project>;
-    remove(projectId: string): Promise<void>;
-    update(projectId: string, updates: Partial<Project>): Promise<Project>;
-    switch(projectId: string): Promise<Project>;
-    openDialog(): Promise<string | null>;
-    onSwitch(callback: (project: Project) => void): () => void;
-  };
-  history: {
-    getSessions(filters?: HistoryGetSessionsPayload): Promise<AgentSession[]>;
-    getSession(sessionId: string): Promise<AgentSession | null>;
-    exportSession(sessionId: string, format: "json" | "markdown"): Promise<string | null>;
-    deleteSession(sessionId: string): Promise<void>;
-  };
-  ai: {
-    getConfig(): Promise<AIConfig>;
-    setKey(apiKey: string): Promise<boolean>;
-    clearKey(): Promise<void>;
-    setModel(model: string): Promise<void>;
-    setEnabled(enabled: boolean): Promise<void>;
-    validateKey(apiKey: string): Promise<boolean>;
-    generateProjectIdentity(projectPath: string): Promise<ProjectIdentity | null>;
-  };
-}
-
 const api: ElectronAPI = {
   // ==========================================
   // Worktree API
@@ -609,14 +244,7 @@ const api: ElectronAPI = {
       return () => ipcRenderer.removeListener(CHANNELS.TERMINAL_EXIT, handler);
     },
 
-    onAgentStateChanged: (
-      callback: (data: {
-        agentId: string;
-        state: string;
-        previousState?: string;
-        timestamp: number;
-      }) => void
-    ) => {
+    onAgentStateChanged: (callback: (data: AgentStateChangePayload) => void) => {
       const handler = (_event: Electron.IpcRendererEvent, data: unknown) => {
         // Type guard
         if (
@@ -804,7 +432,7 @@ const api: ElectronAPI = {
   // AI API
   // ==========================================
   ai: {
-    getConfig: (): Promise<AIConfig> => ipcRenderer.invoke(CHANNELS.AI_GET_CONFIG),
+    getConfig: (): Promise<AIServiceState> => ipcRenderer.invoke(CHANNELS.AI_GET_CONFIG),
 
     setKey: (apiKey: string): Promise<boolean> => ipcRenderer.invoke(CHANNELS.AI_SET_KEY, apiKey),
